+++ sound-mind-studio/tests/test_undo_stack.h
@@ -1,0 +1,18 @@
+#pragma once
+
+#include <QObject>
+
+class UndoStackTest : public QObject {
+    Q_OBJECT
+
+private slots:
+    void freshStackCannotUndoOrRedo();
+    void pushMakesCanUndoTrueAndCanRedoFalse();
+    void undoInvokesTheCommandsOwnUndoCallback();
+    void undoThenRedoInvokesBothCallbacksInOrder();
+    void multipleCommandsUndoAndRedoInReverseThenForwardOrder();
+    void undoWhenEmptyIsANoOp();
+    void redoWhenNothingUndoneIsANoOp();
+    void pushingAfterAnUndoDiscardsTheRedoTail();
+    void clearDiscardsEveryCommandAndResetsAvailability();
+};